+++ src/scherbelberg/_cli/destroy.py
@@ -29,10 +29,18 @@ specific language governing rights and limitations under the License.
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 from asyncio import run
+from logging import ERROR
+import sys
 
 import click
 
-from .._core.cluster import Cluster
+from .._core.cluster import (
+    Cluster,
+    ClusterSchedulerNotFound,
+    ClusterWorkerNotFound,
+    ClusterFirewallNotFound,
+    ClusterNetworkNotFound,
+)
 from .._core.const import PREFIX, TOKENVAR, WAIT
 from .._core.log import configure_log
 
@@ -43,11 +51,29 @@ from .._core.log import configure_log
 
 async def _main(prefix, tokenvar, wait):
 
-    cluster = await Cluster.from_existing(
-        prefix=prefix,
-        tokenvar=tokenvar,
-        wait=wait,
-    )
+    try:
+        cluster = await Cluster.from_existing(
+            prefix=prefix,
+            tokenvar=tokenvar,
+            wait=wait,
+        )
+    except ClusterSchedulerNotFound:
+        click.echo(
+            "Cluster scheduler could not be found. Cluster likely does not exist.",
+            err=True,
+        )
+        sys.exit(1)
+    except (
+        ClusterWorkerNotFound,
+        ClusterFirewallNotFound,
+        ClusterNetworkNotFound,
+    ) as e:
+        click.echo(
+            f"Cluster component missing ({type(e).__name__:s}). Cluster likely needs to be nuked.",
+            err=True,
+        )
+        sys.exit(1)
+
     await cluster.destroy()
 
 
@@ -55,8 +81,9 @@ async def _main(prefix, tokenvar, wait):
 @click.option("-p", "--prefix", default=PREFIX, type=str, show_default=True)
 @click.option("-t", "--tokenvar", default=TOKENVAR, type=str, show_default=True)
 @click.option("-a", "--wait", default=WAIT, type=float, show_default=True)
-def destroy(prefix, tokenvar, wait):
+@click.option("-l", "--log_level", default=ERROR, type=int, show_default=True)
+def destroy(prefix, tokenvar, wait, log_level):
 
-    configure_log()
+    configure_log(log_level)
 
     run(_main(prefix, tokenvar, wait))